Move factories to factory folder
