+++ src/modules/files-storage/domain/factory/file-record.factory.ts
@@ -1,7 +1,7 @@
 import { ObjectId } from '@mikro-orm/mongodb';
 import { EntityId } from '@shared/domain/types';
-import { FileRecord, FileRecordProps, ParentInfo } from './file-record.do';
-import { FileRecordSecurityCheck } from './vo';
+import { FileRecord, FileRecordProps, ParentInfo } from '../file-record.do';
+import { FileRecordSecurityCheck } from '../vo';
 
 export class FileRecordFactory {
 	private static build(fileRecordProps: FileRecordProps, securityCheck: FileRecordSecurityCheck): FileRecord {
