+++ src/modules/files-storage/domain/factory/index.ts
@@ -1,4 +1,6 @@
 export * from './access-url.factory';
 export * from './file-info.factory';
+export * from './file-record.factory';
+export * from './parent-statistic.factory';
 export * from './wopi-access-token.factory';
 export * from './wopi-payload.factory';
