+++ src/modules/files-storage/domain/factory/parent-statistic.factory.ts
@@ -1,4 +1,4 @@
-import { ParentStatistic, ParentStatisticProps } from './vo';
+import { ParentStatistic, ParentStatisticProps } from '../vo';
 
 export class ParentStatisticFactory {
 	public static build(props: ParentStatisticProps): ParentStatistic {
