+++ src/modules/files-storage/domain/index.ts
@@ -3,7 +3,6 @@ export { ErrorType } from './error';
 export * from './factory';
 export * from './file-record.do';
 export * from './interface';
-export { ParentStatisticFactory } from './parent-statistic.factory';
 export { FilesStorageService, PreviewService } from './service';
 export * from './vo';
 export {
